# Conflicts:
#	app/src/com/constellio/app/entities/modules/InstallableSystemModuleExcludedFromSSO.java
#	app/src/com/constellio/app/extensions/AppLayerSystemExtensions.java
#	app/src/com/constellio/app/modules/rm/services/menu/RMRecordsMenuItemServices.java
#	app/src/com/constellio/app/services/migrations/ConstellioEIM.java
#	app/src/com/constellio/app/services/migrations/CoreMigrationCombo.java
#	app/src/com/constellio/app/ui/framework/components/SearchResultDisplay.java
#	app/src/com/constellio/app/ui/framework/components/search/ViewableRecordVOSearchResultTable.java
#	app/src/com/constellio/app/ui/framework/components/viewers/panel/ViewableRecordVOTablePanel.java
#	model/src/com/constellio/model/services/records/cache/RecordsCaches2Impl.java

+++ app/src/com/constellio/app/ui/pages/management/updates/UpdateManagerPresenter.java
@@ -181,11 +181,12 @@ public class UpdateManagerPresenter extends BasePresenter<UpdateManagerView> {
 	}
 
 	public void restartAndReindex(boolean repopulate) {
-		File systemLogFile = getSystemLogFile();
-
-		if (systemLogFile.exists()) {
-			if (!FileUtils.deleteQuietly(systemLogFile)) {
-				view.showErrorMessage($("UpdateManagerViewImpl.error.fileNotDeleted"));
+		if (FoldersLocator.usingAppWrapper()) {
+			File systemLogFile = getSystemLogFile();
+			if (systemLogFile.exists()) {
+				if (!FileUtils.deleteQuietly(systemLogFile)) {
+					view.showErrorMessage($("UpdateManagerViewImpl.error.fileNotDeleted"));
+				}
 			}
 		}
 
@@ -263,7 +264,7 @@ public class UpdateManagerPresenter extends BasePresenter<UpdateManagerView> {
 	private File getSystemLogFile() {
 		File systemLogFile = null;
 
-		File logsFolder = new File(modelLayerFactory.getFoldersLocator().getWrapperInstallationFolder(), "logs");
+		File logsFolder = modelLayerFactory.getFoldersLocator().getLogsFolder();
 		if (logsFolder.exists()) {
 			File tempFile = new File(logsFolder, SYSTEM_LOG_FILE_NAME);
 			if (tempFile.exists()) {
